# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Javari AI Mortgage Rate Monitoring API",
    "version": "1.0.0",
    "description": "Professional mortgage rate monitoring API with real-time data, historical analytics, and email alerts. Built by CR AudioViz AI, LLC.",
    "contact": {
      "name": "CR AudioViz AI Support",
      "email": "support@craudiovizai.com",
      "url": "https://craudiovizai.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://craudiovizai.com/terms"
    }
  },
  "servers": [
    {
      "url": "https://javari-ai.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/mortgage/keys": {
      "get": {
        "tags": ["API Keys"],
        "summary": "List API keys",
        "description": "Retrieves all API keys for the authenticated user.",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of API keys"
          }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Generates a new API key with specified rate limit. Maximum 5 active keys per user.",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "rate_limit"],
                "properties": {
                  "name": {"type": "string", "example": "Production API Key"},
                  "rate_limit": {
                    "type": "integer",
                    "enum": [10, 100, 1000, 10000],
                    "description": "Requests per day"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully. Key is only shown once."
          }
        }
      },
      "delete": {
        "tags": ["API Keys"],
        "summary": "Revoke API key",
        "description": "Revokes an API key by setting is_active to false.",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "key_id",
            "in": "query",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "200": {
            "description": "API key revoked successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for programmatic access. Format: javari_mortgage_xxxxx"
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase JWT token for user authentication"
      }
    }
  }
}